requirements.txt
# Project Dependencies
pandas>=1.3.0
numpy>=1.19.0
scikit-learn>=0.24.0
matplotlib
plotly
feedparser
beautifulsoup4>=4.9.0
requests>=2.25.0
transformers>=4.5.0
torch>=1.9.0
googlesearch-python>=1.2.0
nltk>=3.6.0
